expenzy/requirements.txt
flask~=3.0
gunicorn~=21.2
psycopg2-binary~=2.9
requests~=2.31
